requirements.txt
# Medical Document Orchestration - Python Dependencies

# Core Framework
fastapi
uvicorn

# AI/ML Libraries
google-generativeai
paddleocr

# OCR Engines
pytesseract

# PDF Processing
pdf2image
PyMuPDF

# Image Processing
Pillow
opencv-python

# Data Processing
pandas
numpy
sqlalchemy
pydantic

# Date/Time Processing
dateparser

# Async Processing
aiofiles
asyncpg

# Task Queue
redis
celery

# Utilities
requests
python-multipart
python-dotenv